pyproject.toml
[project]
name = "agentic-ai-digest"
version = "0.1.0"
description = "Agentic AI Digest MVP"
authors = [
    {name = "Vipul Mehra",email = "vipul.m2710@gmail.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "feedparser (>=6.0.11,<7.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "readability-lxml (>=0.8.4.1,<0.9.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "weasyprint (>=66.0,<67.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "sqlite-utils (>=3.38,<4.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "openai (>=1.107.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
